+++ csql/_/api.py
@@ -6,35 +6,42 @@ from .models.overrides import Overrides
 import warnings
 from textwrap import dedent
 
-def Q(
-	sql: Union[str, Callable[[], str]],
-	parameters: Optional[Parameters] = None,
-	dialect: SQLDialect = DefaultDialect,
-	overrides: Optional[Overrides] = None
-) -> Query:
-	""" Make a Query. hi """
-	if callable(sql):
-		warnings.warn(
-			dedent('''
-				Passing a lambda to Q is deprecated! You can now just pass an interpolated string directly:
-					Q(f"select from {blah}")
-			''').strip(),
-			category=DeprecationWarning,
-			stacklevel=2
-		)
-		queryParts = strparsing.getQueryParts(sql())
-	else:
-		queryParts = strparsing.getQueryParts(sql)
+if TYPE_CHECKING:
+	import csql
+	import csql.dialect
 
-	if parameters is not None:
-		warnings.warn(
+def Q(
+	sql: str,
+	dialect: 'csql.dialect.SQLDialect' = DefaultDialect,
+	overrides: Optional['csql.Overrides'] = None
+) -> 'csql.Query':
+	"""
+	Create a :class:`csql.Query`.
+	
+	Usage:
+
+	>>> p = Parameters(created_on=date(2020,1,1))
+	>>> q_cust  = Q(f'select name, customer_type from customers where created_on > {p['created_on']})
+	>>> q_count = Q(f'select customer_type, count(*) from {q1} group by rollup(type)')
+
+	See: TODO: Basic usage
+
+	:param sql: A string with a SQL query. The string is designed to be built with an ``f'f-string'``,
+		so you can interpolate Parameters and other Queries inside in a natural way.
+
+	:param dialect: A default :class:`dialect<csql.dialect.SQLDialect>` to use when building this Query.
+	:param overrides: A default set of :class:`overrides<csql.Overrides>` to use when building this Query.
+
+	"""
+	if callable(sql):
+		raise TypeError(
 			dedent('''
-				Passing parameters to Q as a separate argument is no longer necessary! You can continue to just interpolate params:
-					Q(f"select from {blah} where abc = {p['abc']}")
-			'''),
-			category=DeprecationWarning,
-			stacklevel=2
+				Passing a lambda to Q is no longer supported! You can now just pass an interpolated string directly:
+					Q(f"select from {blah}")
+			''').strip()
 		)
+		
+	queryParts = strparsing.getQueryParts(sql)
 
 	return Query(
 		queryParts=queryParts,