+++ csql/__init__.py
@@ -8,7 +8,7 @@ from ._.models.query import (
 	Parameters,
 	ParameterPlaceholder,
 	RenderedQuery,
-	ParameterList
+	ParameterList as _Deprecated_ParameterList
 )
 
 from ._.models.overrides import Overrides
@@ -21,3 +21,10 @@ __all__ = [
 	'RenderedQuery',
 	'ParameterList'
 ]
+
+def __getattr__(name):
+	if name == 'ParameterList':
+		from warnings import warn
+		warn(f'ParameterList is deprecated. It\'s now just a type alias, and it will be removed from public exports in a future release.', DeprecationWarning)
+		return _Deprecated_ParameterList
+	raise AttributeError(f'module {__name__} has no attribute {name}.')