+++ tests/test_Q_input_strparsing.py
@@ -1,10 +1,10 @@
-from csql import Q, RenderedQuery, ParameterList as PL
+from csql import Q, RenderedQuery
 
 def test_Q_str():
 	q = Q("select 1")
 	assert q.build() == RenderedQuery(
 		sql="select 1",
-		parameters=PL()
+		parameters=()
 	)
 
 def test_Q_dep():